backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.1.1
fastapi==0.115.14
flupy==1.2.2
frozenlist==1.7.0
grpcio==1.73.0
grpcio-tools==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.12
langchain-community==0.3.27
langchain-core==0.3.63
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.2.11
marshmallow==3.26.1
-e git+ssh://git@github.com/felixzc8/mem0.git@155ee650c78eb7181aa3f20e82948955eadc6cb4#egg=mem0ai
multidict==6.5.0
mypy_extensions==1.1.0
mysql-connector-python==8.0.33
numpy==1.26.4
openai==1.88.0
orjson==3.10.18
packaging==24.2
pluggy==1.6.0
portalocker==2.10.1
posthog==5.0.0
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic-settings==2.6.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.12.1
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
StrEnum==0.4.15
tenacity==9.1.2
tidb-vector==0.0.14
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==14.2
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
